ev3dev: add build tags for compilation on other OS
Closes #33
